[lib][y] update

+++ src/mode/model/dialog/ModelAnimationTimelinePanel.cpp
@@ -199,7 +199,7 @@ void ModelAnimationTimelinePanel::on_left_button_up() {
 
 void ModelAnimationTimelinePanel::on_mouse_wheel() {
 	mx = hui::get_event()->m.x;
-	float time_scale_new = min(time_scale * pow(1.1f, hui::get_event()->scroll.y), 1000.0f);
+	float time_scale_new = min(time_scale * (float)pow(1.1f, hui::get_event()->scroll.y), 1000.0f);
 	time_offset += mx * (1.0f / time_scale - 1.0f / time_scale_new);
 	time_scale = time_scale_new;
 	//UpdateTimePos();
